linting

+++ internal/provider/data_sources/catalogue_item_data_source.go
@@ -1,6 +1,3 @@
-// Copyright (c) HashiCorp, Inc.
-// SPDX-License-Identifier: MPL-2.0
-
 package data_sources
 
 import (
@@ -94,7 +91,7 @@ func (d *CatalogueItemDataSource) Read(ctx context.Context, req datasource.ReadR
 		}
 	}
 
-	data.Id = types.Int64Value(int64(matchedCatalogueItem.ID))
+	data.Id = types.Int64Value(matchedCatalogueItem.ID)
 
 	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
 }
