+++ internal/provider/data_sources/category_data_source.go
@@ -1,6 +1,3 @@
-// Copyright (c) HashiCorp, Inc.
-// SPDX-License-Identifier: MPL-2.0
-
 package data_sources
 
 import (
@@ -96,7 +93,7 @@ func (d *CategoryDataSource) Read(ctx context.Context, req datasource.ReadReques
 		}
 	}
 
-	data.Id = types.Int64Value(int64(matchedCategory.CategoryID))
+	data.Id = types.Int64Value(matchedCategory.CategoryID)
 
 	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
 }
