+++ internal/provider/data_sources/form_data_source.go
@@ -1,6 +1,3 @@
-// Copyright (c) HashiCorp, Inc.
-// SPDX-License-Identifier: MPL-2.0
-
 package data_sources
 
 import (
@@ -94,7 +91,7 @@ func (d *FormDataSource) Read(ctx context.Context, req datasource.ReadRequest, r
 		}
 	}
 
-	data.Id = types.Int64Value(int64(matchedForm.FormID))
+	data.Id = types.Int64Value(matchedForm.FormID)
 
 	// Save to Terraform state
 	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
